requirements.txt
aws-cdk.core
aws-cdk.aws-ec2
aws-cdk.aws_iam
aws-cdk.aws_s3_assets
-e .
